sparse: reorganize sparse python codebase
