+++ sparse/runtime.py
@@ -17,7 +17,13 @@ import os
 import numpy as np
 from legate.core import Rect, get_legate_runtime, types
 
-from .config import SparseOpCode, SparseTunable, _supported_dtypes, sparse_ctx
+from .config import (
+    SparseOpCode,
+    SparseProjectionFunctor,
+    SparseTunable,
+    _supported_dtypes,
+    sparse_ctx,
+)
 
 
 class Runtime:
@@ -41,6 +47,7 @@ class Runtime:
                 types.int32,
             )
         )
+        self.dynamic_projection_functor_id = 1
 
         # Register type aliases for all of the common numpy types.
         for np_type, core_type in _supported_dtypes.items():
@@ -65,6 +72,11 @@ class Runtime:
             self.legate_runtime.get_nccl_communicator().initialize(
                 self.num_gpus
             )
+
+    def get_1d_to_2d_functor_id(self, xdim: int, ydim: int, rows: bool):
+        result = self.dynamic_projection_functor_id
+        self.dynamic_projection_functor_id += 1
+        return result + SparseProjectionFunctor.LAST_STATIC_PROJ_FN
 
 
 ctx = sparse_ctx
