+++ sparse/utils.py
@@ -12,9 +12,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import math
 import traceback
 
+import cunumeric
+import numpy
+import pyarrow
+from legate.core import Array, Future, Store, types
 
+from .config import SparseOpCode
+from .runtime import ctx
+
+
+# find_last_user_stacklevel gets the last stack frame index
+# within legate sparse.
 def find_last_user_stacklevel() -> int:
     stacklevel = 1
     for (frame, _) in traceback.walk_stack(None):
@@ -22,3 +33,143 @@ def find_last_user_stacklevel() -> int:
             break
         stacklevel += 1
     return stacklevel
+
+
+# WrappedStore is a helper class that wraps a store into the
+# __legate_data_interface__.
+class WrappedStore:
+    def __init__(self, store: Store):
+        self.store = store
+
+        # Set up the __legate_data_interface__. For some reason
+        # @property isn't working...
+        arrow_type = self.store.type.type
+        if isinstance(arrow_type, numpy.dtype):
+            arrow_type = pyarrow.from_numpy_dtype(arrow_type)
+
+        # We don't have nullable data for the moment
+        # until we support masked arrays
+        array = Array(arrow_type, [None, self.store])
+        data = dict()
+        field = pyarrow.field("cuNumeric Array", arrow_type, nullable=False)
+        data[field] = array
+        self.__legate_data_interface__ = dict()
+        self.__legate_data_interface__["data"] = data
+        self.__legate_data_interface__["version"] = 1
+
+    # @property
+    # def __legate_data_interface__(self)
+    #     if self._legate_data is None:
+    #         # All of our thunks implement the Legate Store interface
+    #         # so we just need to convert our type and stick it in
+    #         # a Legate Array
+    #         arrow_type = pyarrow.from_numpy_dtype(self.store.dtype)
+    #         # We don't have nullable data for the moment
+    #         # until we support masked arrays
+    #         array = Array(arrow_type, [None, self.store])
+    #         self._legate_data = dict()
+    #         self._legate_data["version"] = 1
+    #         data = dict()
+    #         field = pyarrow.field(
+    #             "cuNumeric Array", arrow_type, nullable=False
+    #         )
+    #         data[field] = array
+    #         self._legate_data["data"] = data
+    #     return self._legate_data
+
+
+# store_to_cunumeric_array converts a store to a cuNumeric array.
+def store_to_cunumeric_array(store: Store):
+    return cunumeric.array(WrappedStore(store))
+
+
+# get_store_from_cunumeric_array extracts a store from a cuNumeric array.
+def get_store_from_cunumeric_array(
+    arr: cunumeric.ndarray, allow_future=False
+) -> Store:
+    # TODO (rohany): It's unclear how to actually get stores from the
+    # __legate_data_interface__ for cunumeric arrays. It seems to depend on
+    # whether they are eager/deferred etc. Have to ask Mike and Wonchan about
+    # this.
+    #
+    # data = arr.__legate_data_interface__["data"]
+    # (_, array) = list(data.items())[0]
+    # target = array.stores()[1]
+    # if isinstance(target, Store):
+    #     store = target
+    # else:
+    #     if isinstance(target, cunumeric.eager.EagerArray):
+    #         target = target.to_deferred_array()
+    #     store = target.base
+    #
+    # Because of https://github.com/nv-legate/cunumeric/issues/595, we can't
+    # access stores of cunumeric arrays through the `__legate_data_interface__`
+    # if the stores happen to have a complex type. So, we'll do something
+    # hackier and just reach into the array's thunk and extract the store.
+    target = arr._thunk
+    if isinstance(target, cunumeric.eager.EagerArray):
+        target = target.to_deferred_array()
+    assert isinstance(target, cunumeric.deferred.DeferredArray)
+    store = target.base
+    assert isinstance(store, Store)
+
+    # Our implementation can't handle future backed stores when we use this, as
+    # we expect to be able to partition things up. If we have a future backed
+    # store, create a normal store and issue a copy from the backed store to
+    # the new store.
+    if store.kind == Future and not allow_future:
+        store_copy = ctx.create_store(store.type, shape=store.shape)
+        task = ctx.create_task(SparseOpCode.UPCAST_FUTURE_TO_REGION)
+        task.add_output(store_copy)
+        task.add_input(store)
+        task.add_broadcast(store_copy)
+        task.add_scalar_arg(store.type.size, types.uint64)
+        task.execute()
+        store = store_copy
+    return store
+
+
+# cast_to_store attempts to cast an arbitrary object into a store.
+def cast_to_store(arr):
+    if isinstance(arr, Store):
+        return arr
+    if isinstance(arr, numpy.ndarray):
+        arr = cunumeric.array(arr)
+    if isinstance(arr, cunumeric.ndarray):
+        return get_store_from_cunumeric_array(arr)
+    raise NotImplementedError
+
+
+# cast_arr attempts to cast an arbitrary object into a cunumeric
+# ndarray, with an optional desired type.
+def cast_arr(arr, dtype=None):
+    if isinstance(arr, Store):
+        arr = store_to_cunumeric_array(arr)
+    elif not isinstance(arr, cunumeric.ndarray):
+        arr = cunumeric.array(arr)
+    if dtype is not None:
+        arr = arr.astype(dtype)
+    return arr
+
+
+# require_float64_dtypes is a helper utility to ensure that we only
+# use float64 types for operations without type dispatch until we
+# have type dispatch implemented for all of the operations.
+def require_float64_dtypes(*args):
+    for o in args:
+        if o is None:
+            continue
+        if o.dtype != numpy.float64:
+            raise NotImplementedError(
+                "This operation currently only supports float64 types."
+            )
+
+
+# factor_int decomposes an integer into a close to square grid.
+def factor_int(n):
+    val = math.ceil(math.sqrt(n))
+    val2 = int(n / val)
+    while val2 * val != float(n):
+        val -= 1
+        val2 = int(n / val)
+    return val, val2
